refactor(streaming): extract useSessionInit hook
The session-init concern (everything that constructs and sends
`InitRequest` payloads) moves into one hook:

  - bootstrap effect that fires the initial init on LOADING → connected
  - the "WS closed → reset bootstrap guard + clear placeholder" effect
  - the live re-apply effect that re-sends the config when toggleable
    settings change mid-stream
  - `selectSeed(filename)` for the user-driven seed swap
  - the three coupled refs: `lastSeedRef`, `lastAppliedModelRef`,
    `warmBootstrapSentRef`

`lastAppliedModel` becomes useState-backed (was a ref read at render
time by the lifecycle dispatch payload) — same shape, but it's
actually exposed via the public hook surface now instead of being
read through a ref the lifecycle effects also poked.

`resetSession()` callback replaces the lifecycle effects' direct
ref-pokes (`warmBootstrapSentRef.current = false`,
`lastAppliedModelRef.current = null`) on intentional reconnect and
teardown. `reconnectAfterConnectionLost` calls it too. The lifecycle
effects no longer need either ref.

Provider drops three refs, three effects, one useCallback — down to
519 lines from 627.
